requirements.txt
flask
flask-cors
requests
openai
pinecone-client
python-dotenv
gunicorn
tqdm
python-dotenv